+++ lib/crunchbase/client.rb
@@ -4,6 +4,7 @@ require_relative 'utilities/entity_endpoints'
 require_relative 'utilities/search_endpoints'
 require_relative 'utilities/autocomplete'
 require_relative 'utilities/deleted_entities'
+require_relative 'utilities/downloads'
 
 module Crunchbase
   # API Request
@@ -12,5 +13,6 @@ module Crunchbase
     include Utilities::SearchEndpoints
     include Utilities::Autocomplete
     include Utilities::DeletedEntities
+    include Utilities::Downloads
   end
 end