+++ lib/crunchbase/downloads.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module Crunchbase
+  # Downloads
+  module Downloads
+    autoload :Client, 'crunchbase/downloads/client'
+  end
+end